Remove loader license info
Removed the 'MIT::Loader' license identifier from the PKGBUILD's license array and
deleted the generation of the LICENSE_Loader file.

The loader component is external to the official Burp Suite distribution and
its licensing is separate and potentially subject to EULA restrictions. Including
its license documentation directly within the package license files is unnecessary
and could be misleading.